+++ src/core/settings.h
@@ -153,7 +153,7 @@ struct Values {
     std::string sink_id;
     bool enable_audio_stretching;
     std::string audio_device_id;
-    float sound_volume;
+    u16 sound_volume;
 
     // Camera
     std::array<std::string, Service::CAM::NumCameras> camera_name;
